update FS class
Former-commit-id: 2dc15bb6c506d978ef9415ec680f208a9a4a0203

+++ src/main/java/vsim/utils/FS.java
@@ -24,7 +24,9 @@ import java.io.IOException;
 import java.nio.file.FileVisitOption;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
+import java.nio.file.LinkOption;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.BasicFileAttributes;
@@ -33,7 +35,6 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.stream.Stream;
 
-import vsim.Logger;
 
 
 /**  Basic file system operations. */
@@ -44,13 +45,11 @@ public final class FS {
    *
    * @param file filename
    * @return true if success, false if an error occurs while creating the new file
+   * @throws IOException if an I/O error occurs
    */
-  public static boolean create(Path file) {
-    try {
+  public static void create(Path file) throws IOException {
+    if (!Files.exists(file)) {
       Files.createFile(file);
-      return true;
-    } catch (IOException e) {
-      return false;
     }
   }
 
@@ -59,14 +58,22 @@ public final class FS {
    *
    * @param file file to delete
    * @return true if success, false if an error occurs while deleting
+   * @throws IOException if an I/O error occurs
    */
-  public static boolean delete(Path file) {
-    try {
+  public static void delete(Path file) throws IOException {
+    if (!Files.exists(file)) {
       Files.delete(file);
-      return true;
-    } catch (IOException e) {
-      return false;
     }
+  }
+
+  /**
+   * Converts a String url to Path.
+   *
+   * @param url String url
+   * @return converted url
+   */
+  public static Path toPath(String url) {
+    return Paths.get(url);
   }
 
   /**
@@ -78,6 +85,7 @@ public final class FS {
    * @throws IOException if an I/O error occurs
    */
   public static boolean write(Path file, String text) throws IOException {
+    create(file);
     StandardOpenOption[] opts = new StandardOpenOption[] { WRITE, TRUNCATE_EXISTING };
     Files.write(file, text.getBytes(), opts);
     return true;
@@ -103,38 +111,34 @@ public final class FS {
    *
    * @param dir directory
    * @return list of all files inside the given directory
+   * @throws IOException if an I/O error occurs
    */
-  public static ArrayList<Path> ls(Path dir) {
+  public static ArrayList<Path> ls(Path dir) throws IOException {
     ArrayList<Path> files = new ArrayList<>();
-    try {
-      if (Files.isDirectory(dir)) {
-        HashSet<FileVisitOption> visitOpts = new HashSet<FileVisitOption>(Arrays.asList(FileVisitOption.FOLLOW_LINKS));
-        Files.walkFileTree(dir, visitOpts, Integer.MAX_VALUE, new SimpleFileVisitor<Path>() {
-          @Override
-          public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-            if (attrs.isRegularFile()) {
-              files.add(file);
-            }
-            return FileVisitResult.CONTINUE;
+    if (Files.isDirectory(dir)) {
+      HashSet<FileVisitOption> visitOpts = new HashSet<FileVisitOption>(Arrays.asList(FileVisitOption.FOLLOW_LINKS));
+      Files.walkFileTree(dir, visitOpts, Integer.MAX_VALUE, new SimpleFileVisitor<Path>() {
+        @Override
+        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+          if (attrs.isRegularFile()) {
+            files.add(file);
           }
+          return FileVisitResult.CONTINUE;
+        }
 
-          @Override
-          public FileVisitResult visitFileFailed(Path file, IOException e) throws IOException {
-            return FileVisitResult.SKIP_SUBTREE;
-          }
+        @Override
+        public FileVisitResult visitFileFailed(Path file, IOException e) throws IOException {
+          return FileVisitResult.SKIP_SUBTREE;
+        }
 
-          @Override
-          public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-            return FileVisitResult.CONTINUE;
-          }
-        });
-      }
-    } catch (IOException e) {
-      Logger.warning("something occurred while recursively searching the files in directory");
-    } finally {
-      files.trimToSize();
-      return files;
+        @Override
+        public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
+          return FileVisitResult.CONTINUE;
+        }
+      });
     }
+    files.trimToSize();
+    return files;
   }
 
   /**
@@ -159,8 +163,18 @@ public final class FS {
    * @param path path to verify
    * @return true if the given path is a directory, false if not
    */
-  public static boolean isDirectory(Path path) {
-    return Files.isDirectory(path);
+  public static boolean isDirectory(Path path, LinkOption ...fl) {
+    return Files.isDirectory(path, fl);
+  }
+
+  /**
+   * Returns true if the given path is a hidden path, false if not.
+   *
+   * @param path path to verify
+   * @return true if the given path is a hidden path, false if not
+   */
+  public static boolean isHidden(Path path) {
+    return path.toFile().isHidden();
   }
 
   /**
@@ -181,11 +195,9 @@ public final class FS {
    * @return {@code true} if paths are equal, {@code false} if not
    */
   public static boolean equals(Path a, Path b) {
-    try {
-      return Files.isSameFile(a, b);
-    } catch (IOException e) {
-      return false;
-    }
+    String f1 = a.toFile().getAbsolutePath();
+    String f2 = b.toFile().getAbsolutePath();
+    return f1.equals(f2);
   }
 
   /**
